% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{create_bot_version}
\alias{create_bot_version}
\title{Creates a new version of the bot based on the \code{$LATEST} version}
\usage{
create_bot_version(
  name = "string",
  checksum = "string"
)
}
\arguments{
\item{name}{[required] The name of the bot that you want to create a new version of. The name is case sensitive.}

\item{checksum}{Identifies a specific revision of the \code{$LATEST} version of the bot. If you specify a checksum and the \code{$LATEST} version of the bot has a different checksum, a \code{PreconditionFailedException} exception is returned and Amazon Lex doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the \code{$LATEST} version.}
}
\description{
Creates a new version of the bot based on the \code{$LATEST} version. If the \code{$LATEST} version of this resource hasn't changed since you created the last version, Amazon Lex doesn't create a new version. It returns the last created version.
}
\details{
You can update only the \code{$LATEST} version of the bot. You can't update the numbered versions that you create with the \code{CreateBotVersion} operation.

When you create the first version of a bot, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see versioning-intro.

This operation requires permission for the \code{lex:CreateBotVersion} action.
}
\examples{

}
