% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{delete_bot}
\alias{delete_bot}
\title{Deletes all versions of the bot, including the \code{$LATEST} version}
\usage{
delete_bot(
  name = "string"
)
}
\arguments{
\item{name}{[required] The name of the bot. The name is case sensitive.}
}
\description{
Deletes all versions of the bot, including the \code{$LATEST} version. To delete a specific version of the bot, use the DeleteBotVersion operation.
}
\details{
If a bot has an alias, you can't delete it. Instead, the \code{DeleteBot} operation returns a \code{ResourceInUseException} exception that includes a reference to the alias that refers to the bot. To remove the reference to the bot, delete the alias. If you get the same exception again, delete the referring alias until the \code{DeleteBot} operation is successful.

This operation requires permissions for the \code{lex:DeleteBot} action.
}
\examples{

}
