% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{delete_bot_channel_association}
\alias{delete_bot_channel_association}
\title{Deletes the association between an Amazon Lex bot and a messaging platform}
\usage{
delete_bot_channel_association(
  name = "string",
  botName = "string",
  botAlias = "string"
)
}
\arguments{
\item{name}{[required] The name of the association. The name is case sensitive.}

\item{botName}{[required] The name of the Amazon Lex bot.}

\item{botAlias}{[required] An alias that points to the specific version of the Amazon Lex bot to which this association is being made.}
}
\description{
Deletes the association between an Amazon Lex bot and a messaging platform.
}
\details{
This operation requires permission for the \code{lex:DeleteBotChannelAssociation} action.
}
\examples{

}
