% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{delete_bot_version}
\alias{delete_bot_version}
\title{Deletes a specific version of a bot}
\usage{
delete_bot_version(name, version)
}
\arguments{
\item{name}{[required] The name of the bot.}

\item{version}{[required] The version of the bot to delete. You cannot delete the \code{$LATEST} version of the bot. To delete the \code{$LATEST} version, use the DeleteBot operation.}
}
\description{
Deletes a specific version of a bot. To delete all versions of a bot, use the DeleteBot operation.
}
\details{
This operation requires permissions for the \code{lex:DeleteBotVersion} action.
}
\section{Accepted Parameters}{
\preformatted{delete_bot_version(
  name = "string",
  version = "string"
)
}
}

