% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{delete_intent_version}
\alias{delete_intent_version}
\title{Deletes a specific version of an intent}
\usage{
delete_intent_version(
  name = "string",
  version = "string"
)
}
\arguments{
\item{name}{[required] The name of the intent.}

\item{version}{[required] The version of the intent to delete. You cannot delete the \code{$LATEST} version of the intent. To delete the \code{$LATEST} version, use the DeleteIntent operation.}
}
\description{
Deletes a specific version of an intent. To delete all versions of a intent, use the DeleteIntent operation.
}
\details{
This operation requires permissions for the \code{lex:DeleteIntentVersion} action.
}
\examples{

}
