% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{get_bot}
\alias{get_bot}
\title{Returns metadata information for a specific bot}
\usage{
get_bot(name, versionOrAlias)
}
\arguments{
\item{name}{[required] The name of the bot. The name is case sensitive.}

\item{versionOrAlias}{[required] The version or alias of the bot.}
}
\description{
Returns metadata information for a specific bot. You must provide the bot name and the bot version or alias.
}
\details{
This operation requires permissions for the \code{lex:GetBot} action.
}
\section{Accepted Parameters}{
\preformatted{get_bot(
  name = "string",
  versionOrAlias = "string"
)
}
}

\examples{
# This example shows how to get configuration information for a bot.
\donttest{get_bot(
  name = "DocOrderPizza",
  versionOrAlias = "$LATEST"
)}

}
