% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{get_builtin_slot_types}
\alias{get_builtin_slot_types}
\title{Gets a list of built-in slot types that meet the specified criteria}
\usage{
get_builtin_slot_types(locale = NULL, signatureContains = NULL,
  nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{locale}{A list of locales that the slot type supports.}

\item{signatureContains}{Substring to match in built-in slot type signatures. A slot type will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."}

\item{nextToken}{A pagination token that fetches the next page of slot types. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of slot types, specify the pagination token in the next request.}

\item{maxResults}{The maximum number of slot types to return in the response. The default is 10.}
}
\description{
Gets a list of built-in slot types that meet the specified criteria.
}
\details{
For a list of built-in slot types, see \href{https://developer.amazon.com/public/solutions/alexa/alexa-skills-kit/docs/built-in-intent-ref/slot-type-reference}{Slot Type Reference} in the \emph{Alexa Skills Kit}.

This operation requires permission for the \code{lex:GetBuiltInSlotTypes} action.
}
\section{Accepted Parameters}{
\preformatted{get_builtin_slot_types(
  locale = "en-US"|"en-GB"|"de-DE",
  signatureContains = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

