% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{get_export}
\alias{get_export}
\title{Exports the contents of a Amazon Lex resource in a specified format}
\usage{
get_export(name, version, resourceType, exportType)
}
\arguments{
\item{name}{[required] The name of the bot to export.}

\item{version}{[required] The version of the bot to export.}

\item{resourceType}{[required] The type of resource to export.}

\item{exportType}{[required] The format of the exported data.}
}
\description{
Exports the contents of a Amazon Lex resource in a specified format.
}
\section{Accepted Parameters}{
\preformatted{get_export(
  name = "string",
  version = "string",
  resourceType = "BOT"|"INTENT"|"SLOT_TYPE",
  exportType = "ALEXA_SKILLS_KIT"|"LEX"
)
}
}

