% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{get_import}
\alias{get_import}
\title{Gets information about an import job started with the \code{StartImport} operation}
\usage{
get_import(importId)
}
\arguments{
\item{importId}{The identifier of the import job information to return.}
}
\description{
Gets information about an import job started with the \code{StartImport} operation.
}
\examples{

}
