% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{get_intent_versions}
\alias{get_intent_versions}
\title{Gets information about all of the versions of an intent}
\usage{
get_intent_versions(
  name = "string",
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{name}{[required] The name of the intent for which versions should be returned.}

\item{nextToken}{A pagination token for fetching the next page of intent versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.}

\item{maxResults}{The maximum number of intent versions to return in the response. The default is 10.}
}
\description{
Gets information about all of the versions of an intent.
}
\details{
The \code{GetIntentVersions} operation returns an \code{IntentMetadata} object for each version of an intent. For example, if an intent has three numbered versions, the \code{GetIntentVersions} operation returns four \code{IntentMetadata} objects in the response, one for each numbered version and one for the \code{$LATEST} version.

The \code{GetIntentVersions} operation always returns at least one version, the \code{$LATEST} version.

This operation requires permissions for the \code{lex:GetIntentVersions} action.
}
\examples{

}
