% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{get_slot_types}
\alias{get_slot_types}
\title{Returns slot type information as follows:}
\usage{
get_slot_types(
  nextToken = "string",
  maxResults = 123,
  nameContains = "string"
)
}
\arguments{
\item{nextToken}{A pagination token that fetches the next page of slot types. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch next page of slot types, specify the pagination token in the next request.}

\item{maxResults}{The maximum number of slot types to return in the response. The default is 10.}

\item{nameContains}{Substring to match in slot type names. A slot type will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."}
}
\description{
Returns slot type information as follows:
}
\details{
\itemize{
\item If you specify the \code{nameContains} field, returns the \code{$LATEST} version of all slot types that contain the specified string.
\item If you don't specify the \code{nameContains} field, returns information about the \code{$LATEST} version of all slot types.
}

The operation requires permission for the \code{lex:GetSlotTypes} action.
}
\examples{
# This example shows how to get a list of all of the slot types in your
# account.
get_slot_types(
  maxResults = 10L,
  nextToken = ""
)

}
