% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_package.R
\docType{package}
\name{paws.lexmodelbuildingservice-package}
\alias{paws.lexmodelbuildingservice}
\alias{paws.lexmodelbuildingservice-package}
\title{paws.lexmodelbuildingservice: Amazon Lex Model Building Service}
\description{
Amazon Lex is an AWS service for building
    conversational voice and text interfaces. Use these actions to create,
    update, and delete conversational bots for new and existing client
    applications.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
