% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{get_bot_alias}
\alias{get_bot_alias}
\title{Returns information about an Amazon Lex bot alias}
\usage{
get_bot_alias(name, botName)
}
\arguments{
\item{name}{The name of the bot alias. The name is case sensitive.}

\item{botName}{The name of the bot.}
}
\description{
Returns information about an Amazon Lex bot alias. For more information about aliases, see versioning-aliases.
}
\details{
This operation requires permissions for the \code{lex:GetBotAlias} action.
}
\examples{

}
