% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{get_export}
\alias{get_export}
\title{Exports the contents of a Amazon Lex resource in a specified format}
\usage{
get_export(name, version, resourceType, exportType)
}
\arguments{
\item{name}{The name of the bot to export.}

\item{version}{The version of the bot to export.}

\item{resourceType}{The type of resource to export.}

\item{exportType}{The format of the exported data.}
}
\description{
Exports the contents of a Amazon Lex resource in a specified format.
}
\examples{

}
