% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.lexmodelbuildingservice_operations.R
\name{get_intent}
\alias{get_intent}
\title{Returns information about an intent}
\usage{
get_intent(name, version)
}
\arguments{
\item{name}{The name of the intent. The name is case sensitive.}

\item{version}{The version of the intent.}
}
\description{
Returns information about an intent. In addition to the intent name, you must specify the intent version.
}
\details{
This operation requires permissions to perform the \code{lex:GetIntent} action.
}
\examples{
# This example shows how to get information about an intent.
get_intent(
  version = "$LATEST",
  name = "DocOrderPizza"
)

}
